varios.

+++ server/controllers/filmController.js
@@ -1,20 +1,22 @@
-var express = require('express');
-var Peliculas = require("../models/film");
-var Generos = require("../models/genero");
-var Enlaces = require("../models/enlace");
+let express = require('express');
+let Peliculas = require("../models/film");
+let Generos = require("../models/genero");
+let Enlaces = require("../models/enlace");
 
-var multer = require('multer');
-var path = require('path');
+let utilities = require('../clases/utilities');
 
-var upload = multer({ dest: 'uploads/' });
+let multer = require('multer');
+let path = require('path');
 
-var fs = require("fs");
-var csv = require('fast-csv');
+let upload = multer({ dest: 'uploads/' });
+
+let fs = require("fs");
+let csv = require('fast-csv');
 
 // Middlewares
 const token = require('../middlewares/token');
 
-var router = express.Router();
+let router = express.Router();
 
 // para CORN
 router.use(function (req, res, next) {
@@ -41,19 +43,40 @@ router.route('/peliculas/:id')
     .put(upload.single('portada'),function(req,res){
         Peliculas.findById(req.params.id,function(err,pelicula){
             if(err){
-                res.status(200).json("No existe la pelicula");
+                res.status(500).json({
+                    mensaje: "Hubo un error al tratar de actualizar los datos de la pelicula " + req.body.titulo,
+                    tipo: "danger",
+                    err: err,
+                    visible: true
+                });
             }else{                  
                 
                 var pelicula_update = pelicula;
-                var extension = req.file.originalname.split(".").pop();
+                var extension, foto = "";
 
-                fs.unlink("public/imagenes/films/" + pelicula.foto,  (err) => {
-                    if (err) {
-                        console.log("No se pudo eliminar la imagen:");
-                    } else {
-                        console.log('imagen eliminada');                                
+                try {
+                    extension = req.file.originalname.split(".").pop();
+                    foto = utilities.convertToSlug(req.body.titulo) + "." + extension;
+
+                    if(pelicula.foto == "films450.png"){
+
+                        fs.rename(req.file.path,"public/imagenes/films/" + utilities.convertToSlug(req.body.titulo) + "." + extension, function (err) {
+                            if (err) return console.error(err);
+                        });
+                    
+                    }else{
+                        // Elimino la imagen antes asginada
+                        fs.unlink("public/imagenes/films/" + pelicula.foto,  (err) => {});
+                        // agrego la nueva imagen que se le va a asignar
+                        fs.rename(req.file.path,"public/imagenes/films/" + foto,function (err) {
+                            if (err) return console.error(err);
+                        });
                     }
-                });                                
+
+                } catch (error) {
+                    // mantengo el valor de la foto ya registrada
+                    foto = pelicula.foto;
+                }                                
 
                 pelicula_update.titulo      = typeof req.body.titulo == 'undefined' ? pelicula.titulo : req.body.titulo;
                 pelicula_update.age         = typeof req.body.age == 'undefined' ? pelicula.age : req.body.age;
@@ -63,8 +86,11 @@ router.route('/peliculas/:id')
                 pelicula_update.puntuacion  = typeof req.body.puntuacion == 'undefined' ? pelicula.puntuacion : req.body.puntuacion;
                 pelicula_update.director    = typeof req.body.director == 'undefined' ? pelicula.director : req.body.director;
                 pelicula_update.elenco      = typeof req.body.elenco == 'undefined' ? pelicula.elenco : req.body.elenco;                
+                pelicula_update.is_public   = typeof req.body.is_public == 'undefined' ? pelicula.is_public : req.body.is_public;
+                pelicula_update.slug        = typeof req.body.titulo == 'undefined' ? pelicula.slug : utilities.convertToSlug(req.body.titulo);                                                                                
                 pelicula_update.genero      = typeof req.body.genero == 'undefined' ? pelicula.genero : req.body.genero;
-                pelicula_update.foto        = typeof req.file == 'undefined' ? pelicula.foto : req.body.titulo + "." + extension;
+                pelicula_update.foto        = foto;
+                
 
                 for(var index in req.body.servidor){
                     console.log(typeof req.body._id == 'undefined')
@@ -88,11 +114,6 @@ router.route('/peliculas/:id')
                     }
                 }
 
-                fs.rename(req.file.path,"public/imagenes/films/"+req.body.titulo+"."+extension, function (err) {
-					if (err) return console.error(err);
-					console.log("Imagen enviada a la carpeta de imagenes!");
-				});
-                
                 pelicula_update.save().then(function(us){
                     res.status(200).json({
                         mensaje:"Hemos actualizado los datos de la película " + req.body.titulo + " exitosamente...",
@@ -100,9 +121,10 @@ router.route('/peliculas/:id')
                         visible: true
                     });
                 },function(err){
-                    res.status(200).json({
+                    res.status(500).json({
                         mensaje: "Hubo un error al tratar de actualizar los datos de la pelicula " + req.body.titulo,
                         tipo: "danger",
+                        err: err,
                         visible: true
                     });
                 });
@@ -112,27 +134,29 @@ router.route('/peliculas/:id')
     .delete(function(req,res){
         Peliculas.findById(req.params.id,function(err,pelicula){
             if(err){
-                res.status(200).json("No existe la pelicula");
+                res.status(500).json({
+                    mensaje:"Hubo un error al eliminar los datos de la pelicula " + pelicula.titulo,
+                    tipo: "danger",
+                    visible: true
+                });
             }else{
                 var pelicula_delete = pelicula;
                 
-                Enlaces.remove({pelicula: req.params.id}, function(error){
-                    if(error){
-                        console.log('Error al intentar eliminar el enlace.');
-                    }else{ 
-                        console.log('Enlace eliminado correctamente');
-                    }
-                });
+                Enlaces.remove({pelicula: req.params.id}, function(error){});
 
                 pelicula_delete.remove().then(function(us){
-                    fs.unlink("public/imagenes/films/" + pelicula.foto);                    
+
+                    if(pelicula.foto != "films450.png"){
+                        fs.unlink("public/imagenes/films/" + pelicula.foto);                        
+                    }
+
                     res.status(200).json({
                         mensaje:"Hemos eliminado los datos de la película " + pelicula.titulo + " exitosamente",
                         tipo: "success",
                         visible: true
                     });
                 },function(err){
-                    res.status(200).json({
+                    res.status(500).json({
                         mensaje:"Hubo un error al eliminar los datos de la pelicula " + pelicula.titulo,
                         tipo: "danger",
                         visible: true
@@ -148,32 +172,43 @@ router.route('/peliculas')
         Peliculas.find({})
         .populate("genero")
 		.populate("enlace")
+        .sort({created: -1})
 		.exec(function (err,films) {
 			if (err) {
 				res.json({
-                        mensaje:"Hubo un error al cargar las películas...",
-                        tipo: "danger",
-                        visible: true
-                    }); 
+                    mensaje:"Hubo un error al cargar las películas...",
+                    tipo: "danger",
+                    visible: true
+                }); 
 			}else{
                 res.status(200).json({
-                        mensaje:"Películas cargadas exitosamente...",
-                        tipo: "success",
-                        visible: true,
-                        films: films
-                    });
+                    mensaje:"Películas cargadas exitosamente...",
+                    tipo: "success",
+                    visible: true,
+                    films: films
+                });
 			}
 		});
     })
     .post(upload.single('portada'),function(req,res){
 
-        var extension = req.file.originalname.split(".").pop();
+        var extension, foto = "";
         var longitudEnlace = Object.keys(req.body.servidor).length;
         var idenEnlace = [];
-        // console.log(req.file);
-        // console.log(req.file.path);
-       
-        // console.log(req.body.genero);
+
+        try {
+            extension = req.file.originalname.split(".").pop();
+            foto = utilities.convertToSlug(req.body.titulo) + "." + extension;
+
+
+            fs.rename(req.file.path,"public/imagenes/films/" + foto,function (err) {
+                if (err) return console.error(err);
+            });
+
+        } catch (error) {
+            extension = ".png";
+            foto = "films450" + extension;
+        }
 
         var film = new Peliculas({
             titulo: req.body.titulo,
@@ -184,8 +219,10 @@ router.route('/peliculas')
             puntuacion: req.body.puntuacion,
             director: req.body.director,
             elenco: req.body.elenco,
+            is_public: req.body.is_public,
+            slug: utilities.convertToSlug(req.body.titulo),
             genero: req.body.genero,
-            foto: req.body.titulo + "." + extension
+            foto: foto
         });
         
         for(var i=0; i < longitudEnlace; i++){
@@ -206,20 +243,17 @@ router.route('/peliculas')
 
         film.save(function (err) {
 			if (!err) {
-				fs.rename(req.file.path,"public/imagenes/films/"+req.body.titulo+"."+extension,function (err) {
-					if (err) return console.error(err);
-					// console.log("Succes!");
-					res.status(200).json({
-                        mensaje:"Hemos Registrado los datos de la película " + req.body.titulo + " exitosamente",
-                        tipo: "success",
-                        visible: true
-                    });
-				});
+                res.status(200).json({
+                    mensaje:"Hemos Registrado los datos de la película " + req.body.titulo + " exitosamente",
+                    tipo: "success",
+                    visible: true
+                });
 			}else{
-				res.status(200).json({
+				res.status(500).json({
                     mensaje: "Hubo un error al guardar los datos de la pelicula " + req.body.titulo,
                     tipo: "danger",
-                    visible: true
+                    err: err,
+                    visible: true,
                 });
 			}
 		});
